Convert pthread_join() return code to long not size_t.
- The compiler complains about casting from void* to an int.
- A size_t is the same size as a void*, but it is unsigned.
- A long is signed, and should be the same size on all supported platforms.

+++ src/geopm_pmpi.c
@@ -206,7 +206,7 @@ static int geopm_pmpi_finalize(void)
     if (g_ctl && geopm_env_pmpi_ctl() == GEOPM_PMPI_CTL_PTHREAD) {
         void *return_val;
         err = pthread_join(g_ctl_thread, &return_val);
-        err = err ? err : ((size_t)return_val);
+        err = err ? err : ((long)return_val);
     }
 
     if (!err && g_ctl) {
